+++ src/main/java/ru/ilka/command/admin/SortByHandsWonCommand.java
@@ -32,7 +32,7 @@ public class SortByHandsWonCommand implements ActionCommand {
 
         List<Account> accounts;
         try {
-            accounts = accountLogic.getAllAccounts((int)selfAccount.getAccountId());
+            accounts = accountLogic.getAllAccounts(selfAccount.getAccountId());
             accountLogic.sortByHandsWon(accounts);
         } catch (LogicException e) {
             throw new CommandException("Can't get all accounts " + e);
