+++ src/main/java/ru/ilka/command/admin/SortByProfitCommand.java
@@ -33,7 +33,7 @@ public class SortByProfitCommand implements ActionCommand {
 
         List<Account> accounts;
         try {
-            accounts = accountLogic.getAllAccounts((int)selfAccount.getAccountId());
+            accounts = accountLogic.getAllAccounts(selfAccount.getAccountId());
             accountLogic.sortByProfit(accounts);
         } catch (LogicException e) {
             throw new CommandException("Can't get all accounts " + e);
