+++ src/main/java/ru/ilka/command/guest/RegisterCommand.java
@@ -141,7 +141,7 @@ public class RegisterCommand implements ActionCommand {
                 session.setAttribute(VISITOR_KEY,visitor);
                 session.setAttribute(ACCOUNT_KEY,account);
 
-                ConcurrentHashMap<Long,Account> onlineUsers = (ConcurrentHashMap<Long,Account>) servletContext.getAttribute(ONLINE_USERS_KEY);
+                ConcurrentHashMap<Integer,Account> onlineUsers = (ConcurrentHashMap<Integer,Account>) servletContext.getAttribute(ONLINE_USERS_KEY);
                 if(onlineUsers == null){
                     onlineUsers = new ConcurrentHashMap<>();
                 }
